+++ src/isPromise.test.ts
@@ -5,18 +5,14 @@ import {
 } from "../test/typesDataProvider";
 import { isPromise } from "./isPromise";
 
-describe("isPromise", () => {
-  it("should work as type guard", () => {
-    const data = TYPES_DATA_PROVIDER.promise as AllTypesDataProviderTypes;
-    if (isPromise(data)) {
-      expect(data instanceof Promise).toEqual(true);
-      expectTypeOf(data).toEqualTypeOf<Promise<number>>();
-    }
-  });
+it("should work as type guard", () => {
+  const data = TYPES_DATA_PROVIDER.promise as AllTypesDataProviderTypes;
+  if (isPromise(data)) {
+    expect(data instanceof Promise).toEqual(true);
+  }
+});
 
-  it("should work as type guard in filter", () => {
-    const data = ALL_TYPES_DATA_PROVIDER.filter(isPromise);
-    expect(data.every((c) => c instanceof Promise)).toEqual(true);
-    expectTypeOf(data).toEqualTypeOf<Array<Promise<number>>>();
-  });
+it("should work as type guard in filter", () => {
+  const data = ALL_TYPES_DATA_PROVIDER.filter(isPromise);
+  expect(data.every((c) => c instanceof Promise)).toEqual(true);
 });
